[TMail] refactor: group all AiSuggestReply related classes in one folder
