+++ tmail-backend/tmail-third-party/ai-bot/src/test/java/com/linagora/tmail/common/AiSuggestionMethodTest.java
@@ -23,6 +23,8 @@ import com.linagora.tmail.james.app.MemoryConfiguration;
 import com.linagora.tmail.james.app.MemoryServer;
 import com.linagora.tmail.mailet.conf.AIBaseModule;
 import com.linagora.tmail.module.LinagoraTestJMAPServerModule;
+import com.linagora.tmail.redactionalhelper.AIRedactionalHelper;
+import com.linagora.tmail.redactionalhelper.AIRedactionalHelperForTest;
 import io.restassured.response.Response;
 import io.restassured.specification.RequestSpecification;
 import jakarta.mail.Flags;
