+++ tmail-backend/tmail-third-party/ai-bot/src/test/java/com/linagora/tmail/jmap/aibot/AiBotCapabilityFactoryIntegrationTest.java
@@ -46,6 +46,8 @@ import com.linagora.tmail.james.app.MemoryServer;
 import com.linagora.tmail.mailet.AIRedactionalHelper;
 import com.linagora.tmail.mailet.AIRedactionalHelperForTest;
 import com.linagora.tmail.mailet.conf.AIBaseModule;
+import com.linagora.tmail.redactionalhelper.AIRedactionalHelper;
+import com.linagora.tmail.redactionalhelper.AIRedactionalHelperForTest;
 import com.linagora.tmail.module.LinagoraTestJMAPServerModule;
 
 import io.restassured.specification.RequestSpecification;
