+++ tmail-backend/tmail-third-party/ai-bot/src/test/scala/com/linagora/tmail/AiBotSerializerTest.scala
@@ -18,7 +18,7 @@
 package com.linagora.tmail
 
 import com.linagora.tmail.jmap.aibot.json.AiBotSerializer
-import com.linagora.tmail.jmap.aibot.{AiBotSuggestReplyRequest, AiBotSuggestReplyResponse}
+import com.linagora.tmail.jmap.aibot.redationalHelper.{AiBotSuggestReplyRequest, AiBotSuggestReplyResponse}
 import org.apache.james.jmap.core.{AccountId, Id}
 import org.scalatest.wordspec.AnyWordSpec
 import play.api.libs.json._
